requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
lxml==4.9.2
mypy-extensions==1.0.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
undetected-chromedriver==3.4.6
urllib3==1.26.15
webdriver-manager==3.8.5
websockets==11.0.1
wsproto==1.2.0
fastapi
pydantic
uvicorn
python-multipart
python-dotenv